PostgreSQLと整合するよう修正
